+++ app/view/settings/extraction_settings/quick_draw_settings.py
@@ -70,7 +70,6 @@ class quick_draw_extraction_function(GroupHeaderCardWidget):
         self.half_repeat_spin = SpinBox()
         self.half_repeat_spin.setFixedWidth(WIDTH_SPINBOX)
         self.half_repeat_spin.setMinimum(0)
-        self.half_repeat_spin.setSuffix("次")
         self.half_repeat_spin.setValue(
             readme_settings_async("quick_draw_settings", "half_repeat")
         )
@@ -523,7 +522,6 @@ class quick_draw_basic_animation_settings(GroupHeaderCardWidget):
         self.autoplay_count_spin = SpinBox()
         self.autoplay_count_spin.setFixedWidth(WIDTH_SPINBOX)
         self.autoplay_count_spin.setMinimum(1)
-        self.autoplay_count_spin.setSuffix("次")
         self.autoplay_count_spin.setValue(
             readme_settings_async("quick_draw_settings", "autoplay_count")
         )