+++ app/view/settings/extraction_settings/roll_call_settings.py
@@ -82,7 +82,6 @@ class roll_call_extraction_function(GroupHeaderCardWidget):
         self.half_repeat_spin = SpinBox()
         self.half_repeat_spin.setFixedWidth(WIDTH_SPINBOX)
         self.half_repeat_spin.setMinimum(0)
-        self.half_repeat_spin.setSuffix("次")
         self.half_repeat_spin.setValue(
             readme_settings_async("roll_call_settings", "half_repeat")
         )
@@ -482,7 +481,6 @@ class roll_call_basic_animation_settings(GroupHeaderCardWidget):
         self.autoplay_count_spin = SpinBox()
         self.autoplay_count_spin.setFixedWidth(WIDTH_SPINBOX)
         self.autoplay_count_spin.setMinimum(1)
-        self.autoplay_count_spin.setSuffix("次")
         self.autoplay_count_spin.setValue(
             readme_settings_async("roll_call_settings", "autoplay_count")
         )