+++ app/view/settings/floating_window_management.py
@@ -381,7 +381,7 @@ class floating_window_edge_settings(GroupHeaderCardWidget):
             WIDTH_SPINBOX
         )
         self.floating_window_stick_to_edge_recover_seconds_spinbox.setMinimum(1)
-        self.floating_window_stick_to_edge_recover_seconds_spinbox.setSuffix("秒")
+        self.floating_window_stick_to_edge_recover_seconds_spinbox.setSuffix("s")
         self.floating_window_stick_to_edge_recover_seconds_spinbox.setValue(
             readme_settings_async(
                 "floating_window_management",